The application has become much faster, so disable caching by default
The cache does not have a good invalidation mechanism. Particularly
since it's a common workflow to do something, then go to the web page
and change some labels, delete, etc, this makes it somewhat fragile.

Added an explicit flag to enable caching for scenarios in which the user
is sure they want to use it for something specific. Otherwise, the
query logic is so fast now, that its rarely worth the effort to remember
to provide --clear-cache every single time, to avoid undesirable
behaviour.