Clarify authorize procedure/errors
Rather than simply redirecting the user to the README, inlined the
instructions to go to the developer console for the secrets file into
the rendered error message.

Added a bit of styling to the message to bold certain parts and
differentiate between the pre-formatted part of the error and what
was returned by the libs.